Fix retry mechanism for 400 errors
We were unnecessarily retrying on all 400 errors except 401. This change
should fix the issue.
