+++ compass_sdk/compass.py
@@ -35,6 +35,7 @@ from compass_sdk.constants import (
     DEFAULT_MAX_RETRIES,
     DEFAULT_SLEEP_RETRY_SECONDS,
 )
+from compass_sdk.exceptions import CompassAuthError, CompassClientError, CompassMaxErrorRateExceeded
 from compass_sdk.models import (
     CreateDataSource,
     DataSource,
@@ -48,28 +49,6 @@ from compass_sdk.models import (
 class RetryResult:
     result: Optional[dict] = None
     error: Optional[str] = None
-
-
-class CompassAuthError(Exception):
-    """Exception raised for authentication errors in the Compass client."""
-
-    def __init__(
-        self,
-        message=("CompassAuthError - check your bearer token or username and password."),
-    ):
-        self.message = message
-        super().__init__(self.message)
-
-
-class CompassMaxErrorRateExceeded(Exception):
-    """Exception raised when the error rate exceeds the maximum allowed error rate in the Compass client."""
-
-    def __init__(
-        self,
-        message="The maximum error rate was exceeded. Stopping the insertion process.",
-    ):
-        self.message = message
-        super().__init__(self.message)
 
 
 _DEFAULT_TIMEOUT = 30
@@ -636,7 +615,12 @@ class CompassClient:
         @retry(
             stop=stop_after_attempt(max_retries),
             wait=wait_fixed(sleep_retry_seconds),
-            retry=retry_if_not_exception_type((CompassAuthError, InvalidSchema)),
+            retry=retry_if_not_exception_type(
+                (
+                    CompassClientError,
+                    InvalidSchema,
+                )
+            ),
         )
         def _send_request_with_retry():
             nonlocal error
@@ -667,6 +651,9 @@ class CompassClient:
                 if e.response.status_code == 401:
                     error = "Unauthorized. Please check your username and password."
                     raise CompassAuthError(message=str(e))
+                elif 400 <= e.response.status_code < 500:
+                    error = f"Client error occurred: {e.response.text}"
+                    raise CompassClientError(message=error)
                 else:
                     error = str(e) + " " + e.response.text
                     logger.error(
