+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "compass-sdk"
-version = "0.5.0"
+version = "0.5.1"
 authors = []
 description = "Compass SDK"
